feat: NFC payment service stub
This commits adds a NFC payment service stub in the main profile that
can be enabled on-demand. This is intended to work around a bug in AOSP
where if no payment app is present in the main profile, one cannot even
choose a payment app from the work profile. The stub simply shows up in
the main profile to enable the Settings menu for NFC payment services.

# Conflicts:
#	app/src/main/java/net/typeblog/shelter/ui/SettingsFragment.java
#	app/src/main/java/net/typeblog/shelter/util/SettingsManager.java